feat: add tumbling window and refactor codebase
